Moved menu to own project.
