+++ memecity.engine.ui/Menu/MenuBuilder.h
@@ -1,28 +1,27 @@
 ﻿#ifndef _MENU_BUILDER_H
 #define _MENU_BUILDER_H
 
-#include "Engine/MultimediaManager.h"
-#include "Engine/InputManager.h"
 #include "MenuItem.h"
 #include "Menu.h"
+#include <Engine.h>
 
-namespace menu {
+namespace memecity::engine::ui::menu {
 
 	class MenuBuilder
 	{
 	private:
-		memecity::engine::MultimediaManager& multimedia_manager;
-		memecity::engine::InputManager& input_manager;
+		MultimediaManager& multimedia_manager;
+		InputManager& input_manager;
 		std::unique_ptr<Menu> menu;
 	public:
-		MenuBuilder(memecity::engine::MultimediaManager& multimedia_manager, memecity::engine::InputManager& input_manager) 
+		MenuBuilder(MultimediaManager& multimedia_manager, InputManager& input_manager) 
 		: multimedia_manager(multimedia_manager), input_manager(input_manager) {}
 		MenuBuilder& create_menu(std::string title);
 
-		template<class C, class... Args>
+		template<class... Args>
 		MenuBuilder& with_menu_item(Args&&... args)
 		{
-			menu->create_menu_item<MenuItem>(std::forward<Args>(args)...);
+			menu->create_menu_item(std::forward<Args>(args)...);
 			return *this;
 		}
