+++ memecity.engine.ui/UI.h
@@ -1,0 +1,9 @@
+#ifndef _UI_H
+#define _UI_H
+
+#include "Menu/Menu.h"
+#include "Menu/MenuItem.h"
+#include "Menu/MenuBuilder.h"
+
+#endif
+
