--- memecity.game/Menu/MenuBuilder.cpp
@@ -1,20 +1,0 @@
-﻿#include "MenuBuilder.h"
-
-using namespace menu;
-
-MenuBuilder& MenuBuilder::create_menu(std::string title)
-{
-	menu = std::make_unique<Menu>(multimedia_manager, input_manager, title);
-	return *this;
-}
-
-MenuBuilder& MenuBuilder::with_back_menu_item()
-{
-	menu->create_menu_item<MenuItem>("<-Back", nullptr, [](MenuItem menu_item) {menu_item.unlock(); });
-	return *this;
-}
-
-std::unique_ptr<Menu> MenuBuilder::get_menu()
-{
-	return std::move(menu);
-}
