+++ memecity.game/States/MenuState.h
@@ -2,15 +2,15 @@
 #define _MENU_STATE_H
 
 #include "Engine/State/State.h"
-#include "../Menu/Menu.h"
+#include <UI.h>
 
 class MenuState : public memecity::engine::state::State
 {
 private:
 
-	std::unique_ptr<menu::Menu> advanced_graphics_menu;
-	std::unique_ptr<menu::Menu> settings_menu;
-	std::unique_ptr<menu::Menu> menu;
+	std::unique_ptr<memecity::engine::ui::menu::Menu> advanced_graphics_menu;
+	std::unique_ptr<memecity::engine::ui::menu::Menu> settings_menu;
+	std::unique_ptr<memecity::engine::ui::menu::Menu> menu;
 public:
 	MenuState(memecity::engine::state::StateMachine& sm)
 		: State(sm) {}
