+++ memecity.game/memecity.game.vcxproj
@@ -22,9 +22,6 @@
     <ClInclude Include="Components.h" />
     <ClInclude Include="Components\AIComponent.h" />
     <ClInclude Include="Generate.h" />
-    <ClInclude Include="Menu\Menu.h" />
-    <ClInclude Include="Menu\MenuBuilder.h" />
-    <ClInclude Include="Menu\MenuItem.h" />
     <ClInclude Include="Generate\Strategy\BSP\Fills\Building.h" />
     <ClInclude Include="Generate\Strategy\BSP\Fills\Empty.h" />
     <ClInclude Include="Generate\Strategy\BSP\Fills\Fill.h" />
@@ -65,9 +62,6 @@
     <ClInclude Include="Components\StatsComponent.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Menu\Menu.cpp" />
-    <ClCompile Include="Menu\MenuBuilder.cpp" />
-    <ClCompile Include="Menu\MenuItem.cpp" />
     <ClCompile Include="States\GameState.cpp" />
     <ClCompile Include="States\MenuState.cpp" />
     <ClCompile Include="States\PauseMenuState.cpp" />
@@ -137,10 +131,10 @@
   </ImportGroup>
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <IncludePath>$(SolutionDir)memecity.engine;$(SolutionDir)dependencies\SDL2_mixer-2.0.2\include;$(SolutionDir)dependencies\SDL2_ttf-2.0.14\include;$(SolutionDir)dependencies\SDL2\include;$(SolutionDir)dependencies\SDL2_image-2.0.3\include;$(VC_IncludePath);$(WindowsSDK_IncludePath);$(SolutionDir)memecity.engine.ecs</IncludePath>
+    <IncludePath>$(SolutionDir)memecity.engine.ui;$(SolutionDir)memecity.engine;$(SolutionDir)dependencies\SDL2_mixer-2.0.2\include;$(SolutionDir)dependencies\SDL2_ttf-2.0.14\include;$(SolutionDir)dependencies\SDL2\include;$(SolutionDir)dependencies\SDL2_image-2.0.3\include;$(VC_IncludePath);$(WindowsSDK_IncludePath);$(SolutionDir)memecity.engine.ecs</IncludePath>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-    <IncludePath>$(SolutionDir)memecity.engine;$(SolutionDir)dependencies\SDL2_mixer-2.0.2\include;$(SolutionDir)dependencies\SDL2_ttf-2.0.14\include;$(SolutionDir)dependencies\SDL2\include;$(SolutionDir)dependencies\SDL2_image-2.0.3\include;$(VC_IncludePath);$(WindowsSDK_IncludePath);$(SolutionDir)memecity.engine.ecs</IncludePath>
+    <IncludePath>$(SolutionDir)memecity.engine.ui;$(SolutionDir)memecity.engine;$(SolutionDir)dependencies\SDL2_mixer-2.0.2\include;$(SolutionDir)dependencies\SDL2_ttf-2.0.14\include;$(SolutionDir)dependencies\SDL2\include;$(SolutionDir)dependencies\SDL2_image-2.0.3\include;$(VC_IncludePath);$(WindowsSDK_IncludePath);$(SolutionDir)memecity.engine.ecs</IncludePath>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
@@ -161,7 +155,7 @@
     </ClCompile>
     <Link>
       <AdditionalLibraryDirectories>$(SolutionDir)dependencies\SDL2_mixer-2.0.2\lib\x64;$(SolutionDir)dependencies\SDL2_ttf-2.0.14\lib\x64;$(SolutionDir)dependencies\SDL2_image-2.0.3\lib\x64;$(SolutionDir)x64\Debug;$(SolutionDir)dependencies\SDL2\lib\x64</AdditionalLibraryDirectories>
-      <AdditionalDependencies>memecity.engine.lib;memecity.engine.ecs.lib;SDL2.lib;SDL2main.lib;SDL2_image.lib;SDL2_ttf.lib;SDL2_mixer.lib;kernel32.lib;user32.lib;gdi32.lib;winspool.lib;comdlg32.lib;advapi32.lib;shell32.lib;ole32.lib;oleaut32.lib;uuid.lib;odbc32.lib;odbccp32.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <AdditionalDependencies>memecity.engine.lib;memecity.engine.ecs.lib;memecity.engine.ui.lib;SDL2.lib;SDL2main.lib;SDL2_image.lib;SDL2_ttf.lib;SDL2_mixer.lib;kernel32.lib;user32.lib;gdi32.lib;winspool.lib;comdlg32.lib;advapi32.lib;shell32.lib;ole32.lib;oleaut32.lib;uuid.lib;odbc32.lib;odbccp32.lib;%(AdditionalDependencies)</AdditionalDependencies>
       <SubSystem>Console</SubSystem>
     </Link>
   </ItemDefinitionGroup>
@@ -195,7 +189,7 @@
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
       <SubSystem>Console</SubSystem>
-      <AdditionalDependencies>memecity.engine.lib;memecity.engine.ecs.lib;SDL2.lib;SDL2main.lib;SDL2_image.lib;SDL2_ttf.lib;SDL2_mixer.lib;kernel32.lib;user32.lib;gdi32.lib;winspool.lib;comdlg32.lib;advapi32.lib;shell32.lib;ole32.lib;oleaut32.lib;uuid.lib;odbc32.lib;odbccp32.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <AdditionalDependencies>memecity.engine.lib;memecity.engine.ecs.lib;memecity.engine.ui.lib;SDL2.lib;SDL2main.lib;SDL2_image.lib;SDL2_ttf.lib;SDL2_mixer.lib;kernel32.lib;user32.lib;gdi32.lib;winspool.lib;comdlg32.lib;advapi32.lib;shell32.lib;ole32.lib;oleaut32.lib;uuid.lib;odbc32.lib;odbccp32.lib;%(AdditionalDependencies)</AdditionalDependencies>
       <AdditionalLibraryDirectories>$(SolutionDir)dependencies\SDL2_mixer-2.0.2\lib\x64;$(SolutionDir)dependencies\SDL2_ttf-2.0.14\lib\x64;$(SolutionDir)dependencies\SDL2_image-2.0.3\lib\x64;$(SolutionDir)x64\Release;$(SolutionDir)dependencies\SDL2\lib\x64</AdditionalLibraryDirectories>
     </Link>
   </ItemDefinitionGroup>
