+++ memecity.game/memecity.game.vcxproj.filters
@@ -147,15 +147,6 @@
     <ClInclude Include="Generate.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Menu\Menu.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="Menu\MenuItem.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="Menu\MenuBuilder.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="Generate\Strategy\BSP\Fills\Fill.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -223,15 +214,6 @@
     </ClCompile>
     <ClCompile Include="Systems\MoveSystem.cpp">
       <Filter>Source Files\Systems</Filter>
-    </ClCompile>
-    <ClCompile Include="Menu\Menu.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="Menu\MenuItem.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="Menu\MenuBuilder.cpp">
-      <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="States\GameState.cpp">
       <Filter>Source Files</Filter>
